feat: correção e tratamento de erro no CategoryService

+++ src/main/java/com/movieflix/service/CategoryService.java
@@ -14,20 +14,24 @@ public class CategoryService {
 
     private final CategoryRepository repository;
 
-
-    public List<Category> findAll(){
+    public List<Category> findAll() {
         return repository.findAll();
     }
 
-    public Category saveCategory(Category category){
+    public Category saveCategory(Category category) {
         return repository.save(category);
     }
 
-    public Optional<Category> findById(Long id){
+    public Optional<Category> findById(Long id) {
         return repository.findById(id);
     }
 
-    public void deleteById(Long id){
-        repository.deleteById(id);
+    public void deleteCategory(Long id) {
+        Optional<Category> category = repository.findById(id);
+        if (category.isPresent()) {
+            repository.deleteById(id);
+        } else {
+            throw new RuntimeException("Category not found with id: " + id);
+        }
     }
 }
